added LimbSolverPanel
